refactor: rename shareStationConstants file name to clipboardConstants
