+++ frontend/tests/constants/clipboardConstants.ts
