+++ frontend/tests/homepage.share.station.spec.ts
@@ -5,7 +5,7 @@ import {
   assertToastMessageIsMissing,
 } from "./constants/toasterConstants"
 import { stationWithLocationLatLng, unitedStatesStation } from "./mocks/station"
-import { getClipboardContent } from "./constants/shareStationConstants"
+import { getClipboardContent } from "./constants/clipboardConstants"
 import HomePage from "./pageObjects/HomePage"
 import { homePageUrl } from "./constants/paths"
